coffeeshop/vagrant/coffeeshopsite/requirements.txt
django==3.2.7
django-extensions
django-templatetags
django-countries
djangorestframework
djangorestframework-api-key
psycopg2-binary
sqlalchemy
django-bootstrap4
django-loginas
jupyterlab
django-cors-headers
django-csp
setuptools-rust
django-oauth-toolkit
django-two-factor-auth[phonenumberslite]
django-sslserver
django-mfa2
pep8
